refactor(camera_hub/pairing): move pairing into a module, split logic into general/io/wifi, clean up some code
