+++ camera_hub/src/config.rs
@@ -2,17 +2,20 @@
 //!
 //! SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::DeliveryMonitor;
-use crate::pairing::get_names;
+use crate::pairing::io::get_names;
 use crate::version::camera_version_info;
+use crate::DeliveryMonitor;
 use secluso_client_lib::config::{
-    Heartbeat, HeartbeatRequest, OPCODE_HEARTBEAT_REQUEST, OPCODE_HEARTBEAT_RESPONSE,
-    AddAppRequest, AddAppResponseCommon, AddAppResponseDedicated, OPCODE_ADD_APP_REQUEST, OPCODE_ADD_APP_RESPONSE,
+    AddAppRequest, AddAppResponseCommon, AddAppResponseDedicated, Heartbeat, HeartbeatRequest,
+    OPCODE_ADD_APP_REQUEST, OPCODE_ADD_APP_RESPONSE, OPCODE_HEARTBEAT_REQUEST,
+    OPCODE_HEARTBEAT_RESPONSE,
 };
 use secluso_client_lib::http_client::HttpClient;
-use secluso_client_lib::mls_clients::{NUM_MLS_CLIENTS, MlsClientsCommon, MlsClientsDedicated, CONFIG_DED,
-    NUM_DEDICATED_MLS_CLIENTS, NUM_COMMON_MLS_CLIENTS};
-use secluso_client_lib::mls_client::{MlsClient, ClientType};
+use secluso_client_lib::mls_client::{ClientType, MlsClient};
+use secluso_client_lib::mls_clients::{
+    MlsClientsCommon, MlsClientsDedicated, CONFIG_DED, NUM_COMMON_MLS_CLIENTS,
+    NUM_DEDICATED_MLS_CLIENTS, NUM_MLS_CLIENTS,
+};
 use std::io;
 
 pub fn process_config_command(
@@ -23,11 +26,11 @@ pub fn process_config_command(
     delivery_monitor_opt: Option<&mut DeliveryMonitor>,
     primary_app: bool,
     second_app_already_paired: bool,
-) -> io::Result<Option<MlsClientsDedicated>> {
+) -> anyhow::Result<Option<MlsClientsDedicated>> {
     debug!("Processing config command");
     match clients_ded[CONFIG_DED].decrypt(enc_config_command.to_vec(), true) {
         Ok(command) => {
-            clients_ded[CONFIG_DED].save_group_state().unwrap();
+            clients_ded[CONFIG_DED].save_group_state()?;
             match command[0] {
                 OPCODE_HEARTBEAT_REQUEST => {
                     debug!("Handling heartbeat request");
@@ -39,7 +42,7 @@ pub fn process_config_command(
                         delivery_monitor_opt,
                     )?;
                     Ok(None)
-                },
+                }
                 OPCODE_ADD_APP_REQUEST => {
                     if primary_app {
                         if !second_app_already_paired {
@@ -58,11 +61,11 @@ pub fn process_config_command(
                         error!("Error: Secondary app cannot add other apps!");
                         Ok(None)
                     }
-                },
+                }
                 _ => {
                     error!("Error: Unknown config command opcode!");
                     Ok(None)
-                },
+                }
             }
         }
         Err(e) => {
@@ -98,7 +101,12 @@ fn handle_heartbeat_request(
         );
     }
 
-    send_heartbeat_response(clients_com, clients_ded, heartbeat_request.timestamp, http_client)?;
+    send_heartbeat_response(
+        clients_com,
+        clients_ded,
+        heartbeat_request.timestamp,
+        http_client,
+    )?;
 
     Ok(())
 }
@@ -109,15 +117,19 @@ fn send_heartbeat_response(
     timestamp: u64,
     http_client: &HttpClient,
 ) -> io::Result<()> {
-    let heartbeat = Heartbeat::generate(clients_com, clients_ded, timestamp, camera_version_info()?)?;
+    let heartbeat =
+        Heartbeat::generate(clients_com, clients_ded, timestamp, camera_version_info()?)?;
 
     let mut config_msg = vec![OPCODE_HEARTBEAT_RESPONSE];
     config_msg.extend(bincode::serialize(&heartbeat).unwrap());
 
     let config_msg_enc = clients_ded[CONFIG_DED].encrypt(&config_msg)?;
-    clients_ded[CONFIG_DED].save_group_state().unwrap();
+    clients_ded[CONFIG_DED].save_group_state()?;
 
-    http_client.config_response(&clients_ded[CONFIG_DED].get_group_name().unwrap(), config_msg_enc)?;
+    http_client.config_response(
+        &clients_ded[CONFIG_DED].get_group_name().unwrap(),
+        config_msg_enc,
+    )?;
 
     Ok(())
 }
@@ -127,95 +139,44 @@ fn handle_add_app_request(
     clients_ded: &mut MlsClientsDedicated,
     command_bytes: &[u8],
     http_client: &HttpClient,
-) -> io::Result<Option<MlsClientsDedicated>> {
+) -> anyhow::Result<Option<MlsClientsDedicated>> {
     let add_app_requests: [AddAppRequest; NUM_MLS_CLIENTS] = bincode::deserialize(command_bytes)
         .map_err(|e| io::Error::other(format!("Failed to deserialize add_app msg - {e}")))?;
 
-    let add_app_resps_com: [AddAppResponseCommon; NUM_COMMON_MLS_CLIENTS] = std::array::from_fn(|i| {
-        println!("handle_add_app_request [1]");
-        let camera_key_package = clients_com[i].key_package();
+    let add_app_resps_com: [AddAppResponseCommon; NUM_COMMON_MLS_CLIENTS] =
+        std::array::from_fn(|i| {
+            println!("handle_add_app_request [1]");
+            let camera_key_package = clients_com[i].key_package();
 
-        // FIXME: "app2" is hardcoded.
-        println!("handle_add_app_request [2]");
-        let camera_contact =
-            MlsClient::create_contact("app2", add_app_requests[i].new_app_key_package.clone())
+            // FIXME: "app2" is hardcoded.
+            println!("handle_add_app_request [2]");
+            let camera_contact =
+                MlsClient::create_contact("app2", add_app_requests[i].new_app_key_package.clone())
+                    .unwrap();
+
+            println!("handle_add_app_request [3]");
+            // FIXME: Use a different secret per channel
+            let (welcome_msg_vec, psk_proposal_vec, commit_msg_vec) = clients_com[i]
+                .invite_with_secret(&camera_contact, add_app_requests[i].secret.clone())
                 .unwrap();
 
-        println!("handle_add_app_request [3]");
-        // FIXME: Use a different secret per channel
-        let (welcome_msg_vec, psk_proposal_vec, commit_msg_vec) = clients_com[i]
-            .invite_with_secret(&camera_contact, add_app_requests[i].secret.clone())
-            .unwrap();
+            println!("handle_add_app_request [4]");
+            clients_com[i].save_group_state().unwrap();
 
-        println!("handle_add_app_request [4]");
-        clients_com[i].save_group_state().unwrap();
+            println!("handle_add_app_request [5]");
+            AddAppResponseCommon {
+                camera_key_package,
+                welcome_msg_vec,
+                psk_proposal_vec,
+                commit_msg_vec,
+            }
+        });
 
-        println!("handle_add_app_request [5]");
-        AddAppResponseCommon {
-            camera_key_package,
-            welcome_msg_vec,
-            psk_proposal_vec,
-            commit_msg_vec,
-        }
-    });
-
-    let [(client_l, resp_l), (client_c, resp_c)]: [(MlsClient, AddAppResponseDedicated); NUM_DEDICATED_MLS_CLIENTS] =
-        std::array::from_fn(|i| {
-        // This part of code has a lot in common with initialize_mls_clients() in main.rs
-
-        // Initialize mls_client
-        // FIXME
-        let tag = if i == 0 {
-            "livestream2"
-        } else {
-            "config2"
-        };
-
-        let (camera_name, group_name) = get_names(
-                clients_ded[CONFIG_DED].get_file_dir(), // Could use either of the clients
-                true,
-                format!("camera_{}_name", tag),
-                format!("group_{}_name", tag),
-            );
-        let mut client = MlsClient::new(
-            camera_name,
-            true,
-            clients_ded[CONFIG_DED].get_file_dir(), // Could use either of the clients
-            tag.to_string(),
-            ClientType::Camera,
-        )
-        .expect("MlsClient::new() for returned error.");
-
-        client.create_group(&group_name).unwrap();
-        debug!("Created group.");
-
-        client.save_group_state().unwrap();
-
-        // Now invite
-        let camera_key_package = client.key_package();
-        let app_key_package = add_app_requests[i + NUM_COMMON_MLS_CLIENTS].new_app_key_package.clone();
-        let app_contact = MlsClient::create_contact("app", app_key_package).unwrap();
-        info!("Added contact.");
-
-        let (welcome_msg_vec, _, _) = client
-            .invite_with_secret(&app_contact, add_app_requests[i + NUM_COMMON_MLS_CLIENTS].secret.clone())
-            .inspect_err(|_| {
-                error!("invite() returned error:");
-            }).unwrap();
-        client.save_group_state().unwrap();
-        info!("App invited to the group.");
-
-        // Next, send the shared group name
-        let group_name = client.get_group_name().unwrap();
-
-        let resp = AddAppResponseDedicated {
-            camera_key_package,
-            welcome_msg_vec,
-            group_name,
-        };
-
-        (client, resp)
-    });
+    let [(client_l, resp_l), (client_c, resp_c)]: [(MlsClient, AddAppResponseDedicated);
+        NUM_DEDICATED_MLS_CLIENTS] = [
+        create_client(0, clients_ded, &add_app_requests)?,
+        create_client(1, clients_ded, &add_app_requests)?,
+    ];
 
     let new_clients_ded: MlsClientsDedicated = [client_l, client_c];
     let add_app_resps_ded: [AddAppResponseDedicated; NUM_DEDICATED_MLS_CLIENTS] = [resp_l, resp_c];
@@ -224,13 +185,78 @@ fn handle_add_app_request(
 
     // Send response
     let mut config_msg = vec![OPCODE_ADD_APP_RESPONSE];
-    config_msg.extend(bincode::serialize(&add_app_resp_combined).unwrap());
+    config_msg.extend(bincode::serialize(&add_app_resp_combined)?);
 
     let config_msg_enc = clients_ded[CONFIG_DED].encrypt(&config_msg)?;
     println!("[1]: config_msg_enc len = {:?}", config_msg_enc.len());
-    clients_ded[CONFIG_DED].save_group_state().unwrap();
+    clients_ded[CONFIG_DED].save_group_state()?;
 
-    http_client.config_response(&clients_ded[CONFIG_DED].get_group_name().unwrap(), config_msg_enc)?;
+    http_client.config_response(
+        &clients_ded[CONFIG_DED].get_group_name().unwrap(),
+        config_msg_enc,
+    )?;
 
     Ok(Some(new_clients_ded))
+}
+
+fn create_client(
+    i: usize,
+    clients_ded: &mut MlsClientsDedicated,
+    add_app_requests: &[AddAppRequest; NUM_MLS_CLIENTS],
+) -> anyhow::Result<(MlsClient, AddAppResponseDedicated)> {
+    // This part of code has a lot in common with initialize_mls_clients() in main.rs
+
+    // Initialize mls_client
+    // FIXME
+    let tag = if i == 0 { "livestream2" } else { "config2" };
+
+    let (camera_name, group_name) = get_names(
+        clients_ded[CONFIG_DED].get_file_dir(), // Could use either of the clients
+        true,
+        format!("camera_{}_name", tag),
+        format!("group_{}_name", tag),
+    )?;
+    let mut client = MlsClient::new(
+        camera_name,
+        true,
+        clients_ded[CONFIG_DED].get_file_dir(), // Could use either of the clients
+        tag.to_string(),
+        ClientType::Camera,
+    )
+    .expect("MlsClient::new() for returned error.");
+
+    client.create_group(&group_name)?;
+    debug!("Created group.");
+
+    client.save_group_state()?;
+
+    // Now invite
+    let camera_key_package = client.key_package();
+    let app_key_package = add_app_requests[i + NUM_COMMON_MLS_CLIENTS]
+        .new_app_key_package
+        .clone();
+    let app_contact = MlsClient::create_contact("app", app_key_package)?;
+    info!("Added contact.");
+
+    let (welcome_msg_vec, _, _) = client
+        .invite_with_secret(
+            &app_contact,
+            add_app_requests[i + NUM_COMMON_MLS_CLIENTS].secret.clone(),
+        )
+        .inspect_err(|_| {
+            error!("invite() returned error:");
+        })?;
+    client.save_group_state()?;
+    info!("App invited to the group.");
+
+    // Next, send the shared group name
+    let group_name = client.get_group_name()?;
+
+    let resp = AddAppResponseDedicated {
+        camera_key_package,
+        welcome_msg_vec,
+        group_name,
+    };
+
+    Ok((client, resp))
 }
