+++ camera_hub/src/pairing/mod.rs
@@ -1,0 +1,3 @@
+pub mod wifi;
+pub mod general;
+pub mod io;
